+++ lib/modules/lidarr/routes/add_details.dart
@@ -169,38 +169,6 @@ class _State extends State<LidarrAddDetails> with LunaScrollControllerMixin {
             headers: Database.currentProfileObject.getLidarr()['headers'],
           ),
           ValueListenableBuilder(
-              valueListenable: Database.lunaSeaBox
-                  .listenable(keys: [LidarrDatabaseValue.ADD_MONITORED.key]),
-              builder: (context, box, widget) {
-                return LunaBlock(
-                  title: 'Monitored',
-                  body: const [
-                    TextSpan(text: 'Monitor artist for new releases'),
-                  ],
-                  trailing: LunaSwitch(
-                    value: LidarrDatabaseValue.ADD_MONITORED.data,
-                    onChanged: (value) =>
-                        LidarrDatabaseValue.ADD_MONITORED.put(value),
-                  ),
-                );
-              }),
-          ValueListenableBuilder(
-              valueListenable: Database.lunaSeaBox.listenable(
-                  keys: [LidarrDatabaseValue.ADD_ALBUM_FOLDERS.key]),
-              builder: (context, box, widget) {
-                return LunaBlock(
-                  title: 'Use Album Folders',
-                  body: const [
-                    TextSpan(text: 'Sort tracks into album folders'),
-                  ],
-                  trailing: LunaSwitch(
-                    value: LidarrDatabaseValue.ADD_ALBUM_FOLDERS.data,
-                    onChanged: (value) =>
-                        LidarrDatabaseValue.ADD_ALBUM_FOLDERS.put(value),
-                  ),
-                );
-              }),
-          ValueListenableBuilder(
             valueListenable: Database.lunaSeaBox
                 .listenable(keys: [LidarrDatabaseValue.ADD_ROOT_FOLDER.key]),
             builder: (context, box, widget) {
@@ -221,6 +189,24 @@ class _State extends State<LidarrAddDetails> with LunaScrollControllerMixin {
               );
             },
           ),
+          LidarrDatabaseValue.ADD_MONITORED_STATUS.listen(
+              builder: (context, box, _) {
+            LidarrDatabaseValue _db = LidarrDatabaseValue.ADD_MONITORED_STATUS;
+            LidarrMonitorStatus _status =
+                LidarrMonitorStatus.ALL.fromKey(_db.data);
+            _status ??= LidarrMonitorStatus.ALL;
+
+            return LunaBlock(
+              title: 'Monitor',
+              trailing: const LunaIconButton.arrow(),
+              body: [TextSpan(text: _status.readable)],
+              onTap: () async {
+                Tuple2<bool, LidarrMonitorStatus> _result =
+                    await LidarrDialogs().selectMonitoringOption(context);
+                if (_result.item1) _db.put(_result.item2.key);
+              },
+            );
+          }),
           ValueListenableBuilder(
             valueListenable: Database.lunaSeaBox.listenable(
                 keys: [LidarrDatabaseValue.ADD_QUALITY_PROFILE.key]),
@@ -274,8 +260,8 @@ class _State extends State<LidarrAddDetails> with LunaScrollControllerMixin {
           LidarrDatabaseValue.ADD_QUALITY_PROFILE.data,
           LidarrDatabaseValue.ADD_ROOT_FOLDER.data,
           LidarrDatabaseValue.ADD_METADATA_PROFILE.data,
-          LidarrDatabaseValue.ADD_MONITORED.data ?? true,
-          LidarrDatabaseValue.ADD_ALBUM_FOLDERS.data ?? true,
+          LidarrMonitorStatus.ALL
+              .fromKey(LidarrDatabaseValue.ADD_MONITORED_STATUS.data),
           search: search,
         )
         .then((id) => Navigator.of(context)